abstract manifest getters
